requirements copy.txt
pypdf==4.0.1
python-dotenv==1.0.1
flask==3.0.2
openai==1.12.0
flask-cors==4.0.0
gcsa==2.2.0
python-dateutil==2.8.2 
gunicorn==21.2.0 
authlib==1.0.0